+++ src/actions.rs
@@ -5,12 +5,16 @@ pub mod insert;
 pub mod motion;
 pub mod search;
 pub mod select;
+#[cfg(feature = "system-editor")]
+pub mod system_editor;
 use crate::state::selection::Selection;
 use crate::{EditorMode, EditorState};
 use cpaste::PasteOverSelection;
 use delete::DeleteToEndOfLine;
 use enum_dispatch::enum_dispatch;
 use motion::{MoveToFirstRow, MoveToLastRow};
+#[cfg(feature = "system-editor")]
+pub use system_editor::OpenSystemEditor;
 
 pub use self::cpaste::{CopyLine, CopySelection, Paste};
 pub use self::delete::{
@@ -86,6 +90,8 @@ pub enum Action {
     SelectCurrentSearch(SelectCurrentSearch),
     AppendCharToSearch(AppendCharToSearch),
     RemoveCharFromSearch(RemoveCharFromSearch),
+    #[cfg(feature = "system-editor")]
+    OpenSystemEditor(OpenSystemEditor),
 }
 
 #[enum_dispatch]