+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.10.3 - ?
+===================
+- Add `OpenSystemEditor` which opens a system editor.
+    - Invoked with `ctrl+e` in vim mode.
+    - Invoked with `alt+e` in emacs mode.
+
 Released
 --------